net: New iface_stat module to monitor persistent network stats.
Adding a new module that tracks the number of bytes/packets transfered by a network interface, even after the interface has been removed.
This is relevant to track the total data usage in mobile devices whose interfaces are added and removed quite frequently (WiFi, Bluetooth, 3G,..).

Monitoring is done only for devices that are configured with a valid IP address (check to exclude virtual/loopback/tunnel interfaces).

Change-Id: I8ac642af1990433ebd0784e8dbd72bf0714b5bf6
Signed-off-by: Ashish Sharma <ashishsharma@google.com>

Included a little modification for the out-of-date HD2 .32 kernel.